♻️ refactor(work): rename fileWorkRegistration to workRegistration
