+++ apps/server/src/services/agentRuntime/workRegistration.ts
